base64ct: refactor into modules
Refactors the single-file `lib.rs` implementation into separate modules
for encoding/decoding and specific "flavors" of Base64.

This is in preparation for adding another new Base64 flavor
(the "libcrypt" flavor used by bcrypt and SHA-crypt password hashes)
